Give the size options one type and the two bare integers their enums (#135)
Items 6 and 7 of #125.

**Twenty size members are ulong rather than nuint.** The same native size_t was
ulong on some members and nuint on others: WriteBufferSize against
ArenaBlockSize, Cache.CreateLru against WriteBufferManager.Create,
HyperClockCacheOptions taking ulong and exposing nuint. nuint is also
platform-dependent in a public API, which is not a property a size option
should have. The conversion sits at the P/Invoke boundary as
checked((nuint)value), so on a 32-bit process a value above uint.MaxValue
throws rather than truncating to something smaller than asked for.

The generated layer is untouched and stays nuint, because that is what
size_t is. This was a wrapper problem, not a generator one.

**BlockBasedTableOptions.Checksum is the ChecksumType enum.** It was an int on
the grounds that the C API does not publish the values. It does not, but the
header does: ChecksumType in include/rocksdb/table.h at the pinned version, five
members, and the numbering cannot drift because the value is written into every
SST file and read back by whichever version opens it later.

Checked against the running library as well as the header. RocksDb accepts
exactly 0 to 4 and rejects anything outside, naming ChecksumType when it does; a
fresh options object reports 4; and RocksDb's own option parser accepts exactly
the five names the members are called after, and rejects an invented one. Three
tests pin those.

**DbOptions.VerifyOutputFlags is a [Flags] enum.** Described as flags RocksDb
does not expose, and it does: VerifyOutputFlags in
include/rocksdb/advanced_options.h, seven members. The interesting part is the
shape, which nothing said: the bits come in two groups, three saying what to
verify and two saying which compactions to verify it on, and a value needs one
from each group to do anything. That is why setting a single bit had no
observable effect. The C API takes and returns int against RocksDb's uint32_t,
which is why All arrives as -1, and why the old int property accepted -1 as a
value: that was asking for everything.

Not done, so #125 stays open: the ticker and histogram ids that GetTickerCount
and GetHistogramData take as bare numbers. Measured in statistics.h at the
pinned version, those are 263 and 80 members with almost entirely implicit,
positional values, and they gain and lose entries between releases. Hand-writing
343 members that renumber when RocksDb reorders one is the wrong shape; the
generator already fetches c.h at the pinned tag and should fetch statistics.h
the same way. That is its own change.

The changelog gains eleven rows, covering these three and the five items merged
in #134, which never made it into the migration table the README points at. The
row for WriteBatchWithIndex.DeleteRange described it as throwing and obsolete
and is now correct about it being gone.

1054 tests pass on net8.0, net9.0 and net10.0. Build clean under -warnaserror
from a clean build.

+++ RocksDb.Net.Tests/DbOptionsSweepTests.cs
@@ -58,7 +58,7 @@ public class DbOptionsSweepTests
         using var opts = new DbOptions
         {
             ArenaBlockSize = 64 * 1024,
-            HardPendingCompactionBytesLimit = (nuint)(1UL << 30),
+            HardPendingCompactionBytesLimit = (ulong)(1UL << 30),
             InplaceUpdateNumLocks = 500,
             LogFileTimeToRoll = 3600,
             ManifestPreallocationSize = 8 * 1024 * 1024,
@@ -66,19 +66,19 @@ public class DbOptionsSweepTests
             MaxSuccessiveMerges = 12,
             MemtableHugePageSize = 2 * 1024 * 1024,
             RecycleLogFileNum = 4,
-            SoftPendingCompactionBytesLimit = (nuint)(1UL << 28),
+            SoftPendingCompactionBytesLimit = (ulong)(1UL << 28),
         };
 
-        Assert.Equal((nuint)(64 * 1024), opts.ArenaBlockSize);
-        Assert.Equal((nuint)(1UL << 30), opts.HardPendingCompactionBytesLimit);
-        Assert.Equal((nuint)500, opts.InplaceUpdateNumLocks);
-        Assert.Equal((nuint)3600, opts.LogFileTimeToRoll);
-        Assert.Equal((nuint)(8 * 1024 * 1024), opts.ManifestPreallocationSize);
-        Assert.Equal((nuint)(32 * 1024 * 1024), opts.MaxManifestFileSize);
-        Assert.Equal((nuint)12, opts.MaxSuccessiveMerges);
-        Assert.Equal((nuint)(2 * 1024 * 1024), opts.MemtableHugePageSize);
-        Assert.Equal((nuint)4, opts.RecycleLogFileNum);
-        Assert.Equal((nuint)(1UL << 28), opts.SoftPendingCompactionBytesLimit);
+        Assert.Equal((ulong)(64 * 1024), opts.ArenaBlockSize);
+        Assert.Equal((ulong)(1UL << 30), opts.HardPendingCompactionBytesLimit);
+        Assert.Equal((ulong)500, opts.InplaceUpdateNumLocks);
+        Assert.Equal((ulong)3600, opts.LogFileTimeToRoll);
+        Assert.Equal((ulong)(8 * 1024 * 1024), opts.ManifestPreallocationSize);
+        Assert.Equal((ulong)(32 * 1024 * 1024), opts.MaxManifestFileSize);
+        Assert.Equal((ulong)12, opts.MaxSuccessiveMerges);
+        Assert.Equal((ulong)(2 * 1024 * 1024), opts.MemtableHugePageSize);
+        Assert.Equal((ulong)4, opts.RecycleLogFileNum);
+        Assert.Equal((ulong)(1UL << 28), opts.SoftPendingCompactionBytesLimit);
     }
 
     [Fact]
@@ -257,7 +257,7 @@ public class DbOptionsSweepTests
             CompactionPri = CompactionPri.RoundRobin,
             DeleteObsoleteFilesPeriodMicros = 6_000_000,
             EnablePipelinedWrite = true,
-            HardPendingCompactionBytesLimit = (nuint)(1UL << 30),
+            HardPendingCompactionBytesLimit = (ulong)(1UL << 30),
             IsFdCloseOnExec = true,
             MaxFileOpeningThreads = 4,
             MaxManifestFileSize = 16 * 1024 * 1024,
@@ -265,7 +265,7 @@ public class DbOptionsSweepTests
             OptimizeFiltersForHits = true,
             RecycleLogFileNum = 2,
             ReportBgIoStats = true,
-            SoftPendingCompactionBytesLimit = (nuint)(1UL << 28),
+            SoftPendingCompactionBytesLimit = (ulong)(1UL << 28),
             StatisticsLevel = StatsLevel.ExceptDetailedTimers,
             StatsPersistPeriodSec = 60,
             TableCacheNumShardBits = 4,
